made refactoring, added executing program. final
