+++ data/stats.csv
@@ -1,0 +1,8 @@
+1,37.86729431152344
+2,24.54872226715088
+3,20.36780571937561
+4,21.03993844985962
+5,21.382314682006836
+6,20.62287187576294
+7,19.284966468811035
+8,19.6474666595459
